+++ mypy_json_report.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import argparse
+import enum
 import json
 import sys
 from collections import Counter, defaultdict
@@ -19,8 +21,54 @@ from dataclasses import dataclass
 from typing import Counter as CounterType, Dict, Iterator
 
 
+class ErrorCodes(enum.IntEnum):
+    DEPRECATED = 1
+
+
 def main() -> None:
-    print(produce_errors_report(sys.stdin))
+    """
+    The primary entrypoint of the program.
+
+    Parses the CLI flags, and delegates to other functions as appropriate.
+    For details of how to invoke the program, call it with `--help`.
+    """
+    parser = argparse.ArgumentParser()
+    subparsers = parser.add_subparsers(title="subcommand")
+
+    parser.set_defaults(func=_no_command)
+
+    parse_parser = subparsers.add_parser(
+        "parse", help="Transform Mypy output into JSON."
+    )
+    parse_parser.add_argument(
+        "-i",
+        "--indentation",
+        type=int,
+        default=2,
+        help="Number of spaces to indent JSON output.",
+    )
+
+    parse_parser.set_defaults(func=_parse_command)
+
+    parsed = parser.parse_args()
+    parsed.func(parsed)
+
+
+def _parse_command(args: argparse.Namespace) -> None:
+    """Handle the `parse` command."""
+    errors = parse_errors_report(sys.stdin)
+    error_json = json.dumps(errors, sort_keys=True, indent=args.indentation)
+    print(error_json)
+
+
+def _no_command(args: argparse.Namespace) -> None:
+    """
+    Handle the lack of an explicit command.
+
+    This will be hit when the program is called without arguments.
+    """
+    print("A subcommand is required. Pass --help for usage info.")
+    sys.exit(ErrorCodes.DEPRECATED)
 
 
 @dataclass(frozen=True)
@@ -29,12 +77,12 @@ class MypyError:
     message: str
 
 
-def produce_errors_report(input_lines: Iterator[str]) -> str:
-    """Given lines from mypy's output, return a JSON summary of error frequencies by file."""
+def parse_errors_report(input_lines: Iterator[str]) -> Dict[str, Dict[str, int]]:
+    """Given lines from mypy's output, return a summary of error frequencies by file."""
     errors = _extract_errors(input_lines)
     error_frequencies = _count_errors(errors)
     structured_errors = _structure_errors(error_frequencies)
-    return json.dumps(structured_errors, sort_keys=True, indent=2)
+    return structured_errors
 
 
 def _extract_errors(lines: Iterator[str]) -> Iterator[MypyError]: