+++ tests/test_mypy_json_report.py
@@ -1,7 +1,6 @@
-import json
 from io import StringIO
 
-from mypy_json_report import produce_errors_report
+from mypy_json_report import parse_errors_report
 
 
 EXAMPLE_MYPY_STDOUT = """\
@@ -11,10 +10,10 @@ mypy_json_report.py:68: error: Call to untyped function "main" in typed context
 Found 2 errors in 1 file (checked 3 source files)"""
 
 
-def test_report() -> None:
-    report = produce_errors_report(StringIO(EXAMPLE_MYPY_STDOUT))
+def test_parse_errors_report() -> None:
+    report = parse_errors_report(StringIO(EXAMPLE_MYPY_STDOUT))
 
-    assert json.loads(report) == {
+    assert report == {
         "mypy_json_report.py": {
             'Call to untyped function "main" in typed context': 1,
             "Function is missing a return type annotation": 1,